Preserving more binary data
I do not tear things down and rip them apart to hack when I write my code. Hackers (as a whole bunch) are an opinionated group of passionate individuals who have a responsibility to be the first line of defense when something like this happens.

Contextual reference point is here: https://github.com/emsec/ChameleonMini/issues/271#issuecomment-679546093. The reverse engineered RFIDTools APK (off of Play Store) is much more authentic to a typical Android application. Uploading this data to preserve it and make that point clearly. The zipped JADX archive is too large, and hence GitHub denies me the ability to upload it directly.